[WPE][GTK][WTF] Implement SystemTracing.h with Sysprof https://bugs.webkit.org/show_bug.cgi?id=275314
Reviewed by NOBODY (OOPS!).

Implement the tracing macros provided by SystemTracing.h using the
Sysprof collector API.

Sysprof collector operates by passing a control fd as an environment
variable. This env var is set when running any executable with the
sysprof-cli utility, or through Sysprof's main interface.

The control fd needs to be propagated to child processes properly. For
the case of run-minibrowser script - a Python script that spawns the
actual MiniBrowser executable - and for the subprocess launcher. To do
that, duplicate the control fd and pass it as an environment variable
to subprocesses well. The Python script neatly supports that, but the
subprocess launcher needed some more manual mangling to get that done.

This initial work covers all pre-existing tracing groups implemented
on kdebug.

This patch does not introduce new tracing points.

There are limitations with the tracing macros though, for example,
they are most meaningful when there's a single WebKit view running.
Once more views are created, they all output into the same tracing
group. In the future, this can be fixed by using the subprocesses'
pids as part of the tracing name, but right now it's most useless
as the sandbox makes all subprocesses have pid 3.

Another limitation is that tracing only truly works when disabling
the sandboxing of subprocesses. Working around it might require some
reengineering of the subprocess launching mechanism, so this is left
out of this patch as well.

To run WebKitGTK with Sysprof tracing, run the following command in
the WebKit source directory:

```
$ sysprof-cli --force -e WEBKIT_DISABLE_SANDBOX_THIS_IS_DANGEROUS=1 -- Tools/Scripts/run-minibrowser --gtk
```

For WPE WebKit, run the following command:

```
$ sysprof-cli --force -e WEBKIT_DISABLE_SANDBOX_THIS_IS_DANGEROUS=1 -- Tools/Scripts/run-minibrowser --wpe
```

It is also possible to get GNOME Shell marks by passing `--gnome-shell`
to the sysprof-cli command.

* Source/JavaScriptCore/jit/JITPlan.cpp:
(JSC::JITPlan::compileInThread):
* Source/WTF/wtf/PlatformGTK.cmake:
* Source/WTF/wtf/PlatformWPE.cmake:
* Source/WTF/wtf/SystemTracing.h:
(WTF::tracePoint):
* Source/WTF/wtf/glib/SysprofAnnotator.h: Added.
* Source/WebCore/dom/Document.cpp:
(WebCore::Document::setReadyState):
* Source/WebKit/GPUProcess/glib/GPUProcessMainGLib.cpp:
(WebKit::GPUProcessMain):
* Source/WebKit/UIProcess/API/glib/WebKitInitialize.cpp:
(WebKit::webkitInitialize):
* Source/WebKit/UIProcess/Launcher/glib/ProcessLauncherGLib.cpp:
(WebKit::ProcessLauncher::launchProcess):
* Source/WebKit/WebProcess/gtk/WebProcessMainGtk.cpp:
* Tools/Scripts/webkitpy/common/system/abstractexecutive.py:
(AbstractExecutive.run_command):
* Tools/Scripts/webkitpy/common/system/executive.py:
(Executive.run_command):
* Tools/Scripts/webkitpy/common/system/executive_mock.py:
(MockExecutive.run_command):
* Tools/Scripts/webkitpy/port/gtk.py:
(GtkPort.run_minibrowser):